+++ constants.py
@@ -1,12 +1,15 @@
 from enum import Enum
 
+
 class Role(str, Enum):
     ADMIN = "admin"
     PARENT = "parent"
 
+
 class SecretKeys:
     GCP = "gcp"
     GCP_SA = "gcp_service_account"
+
 
 class GcpSecretFields:
     PHOTOS_ROOT = "drive_photos_root_folder_id"